admin_plugin/dev/requirements.txt
dataiku-api-client==9.0.0